Created fading.py and growing.py to help break up animation.creation.py
